pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codesight"
version = "0.1.0"
description = "CodeSight - GitLab AI code error detector and fixer"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn==0.35.0",
  "python-dotenv==1.1.1",
  "requests==2.32.4",
  "pydantic==2.11.7"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
